+++ packages/gamut/src/GridForm/GridFormButtons/index.tsx
@@ -3,39 +3,21 @@ import React, { ComponentProps } from 'react';
 import { Box } from '../../Box';
 import { CTAButton, FillButton, TextButton } from '../../Button';
 import { ButtonProps } from '../../Button/shared';
-import {
-  ButtonDeprecated,
-  ButtonDeprecatedProps,
-} from '../../ButtonDeprecated';
 import { Column } from '../../Layout';
 import { VisualTheme } from '../../theming/VisualTheme';
 
 export type GridFormButtonsPosition = keyof typeof positions;
 
 export type SubmitButtonType = 'cta' | 'fill';
-export type SubmitButtonDeprecatedType = 'business';
 
-type GridFormBase = {
+export type GridFormSubmitProps = {
   contents: React.ReactNode;
   position?: GridFormButtonsPosition;
   size: ComponentProps<typeof Column>['size'];
-  disabled?: ButtonDeprecatedProps['disabled'];
+  disabled?: ButtonProps['disabled'];
   mode?: VisualTheme;
-};
-
-type GridFormButtonSubmitPropsDeprecated = GridFormBase & {
-  type?: SubmitButtonDeprecatedType;
-  theme?: ButtonDeprecatedProps['theme'];
-  outline?: ButtonDeprecatedProps['outline'];
-};
-
-type GridFormSubmitPropsStandard = GridFormBase & {
   type?: SubmitButtonType;
 };
-
-export type GridFormSubmitProps =
-  | GridFormButtonSubmitPropsDeprecated
-  | GridFormSubmitPropsStandard;
 
 export type GridFormCancelButtonProps = {
   children: React.ReactNode;
@@ -65,25 +47,7 @@ export const GridFormButtons: React.FC<
             {props.contents}
           </CTAButton>
         );
-      case 'business':
-        /**
-         * There are current designs that currently rely on the deprecated button.
-         * Primarily business components such as /WorkerSupportApplication/index.tsx,
-         * /PlanInvitationBulkForm/PlanInvitationBulkForm.tsx and /PlanInvitationForm/PlanInvitationForm.tsx
-         * currently using brand-blue variant of the deprecated button. With the later two also using the also
-         * using the outline button. Once work is finished for the colorMode changes for buttons the deprecated button
-         * can be removed.
-         */
-        return (
-          <ButtonDeprecated
-            disabled={props.disabled}
-            outline={props.outline}
-            theme={props.theme}
-            type="submit"
-          >
-            {props.contents}
-          </ButtonDeprecated>
-        );
+
       default:
         return (
           <FillButton type="submit" mode={props.mode} disabled={props.disabled}>